requirements.txt
Flask
python-dotenv
google-cloud-dialogflow
requests